requirements.txt
imgurpython~=1.1.7
pillow~=8.4.0
pytest~=6.2.5
asyncpraw~=7.5.0
requests~=2.26.0
